chore: apply machine-applicable clippy fixes across the workspace
cargo clippy --fix output only: unused imports, needless casts, redundant
closures, collapsible ifs, ?-operator shortcuts, contains() over
iter().any(), and the like. No behavior change. std::fs imports that only
tests use moved into their #[cfg(test)] modules.

+++ crates/grokhub-app/src/markdown.rs
@@ -3,7 +3,7 @@ use grokhub_core::{bubble_max_width, TEXT_FILE_CAP};
 
 /// Paint/layout prefix. Stream buffers may hold `IMAGE_FILE_CAP`; laying that out freezes Chat.
 pub(crate) fn display_text(text: &str) -> &str {
-    let cap = TEXT_FILE_CAP as usize;
+    let cap = TEXT_FILE_CAP;
     if text.len() <= cap {
         return text;
     }
